Separate AST into parts
