+++ Sources/Optimizer/Optimizer.swift
@@ -11,268 +11,4 @@ import AST
 public struct Optimizer: ASTPass {
 
   public init() {}
-
-  public func process(topLevelModule: TopLevelModule, passContext: ASTPassContext) -> ASTPassResult<TopLevelModule> {
-    return ASTPassResult(element: topLevelModule, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(topLevelDeclaration: TopLevelDeclaration, passContext: ASTPassContext) -> ASTPassResult<TopLevelDeclaration> {
-    return ASTPassResult(element: topLevelDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(contractDeclaration: ContractDeclaration, passContext: ASTPassContext) -> ASTPassResult<ContractDeclaration> {
-    return ASTPassResult(element: contractDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(contractBehaviorDeclaration: ContractBehaviorDeclaration, passContext: ASTPassContext) -> ASTPassResult<ContractBehaviorDeclaration> {
-    return ASTPassResult(element: contractBehaviorDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(contractBehaviorMember: ContractBehaviorMember, passContext: ASTPassContext) -> ASTPassResult<ContractBehaviorMember> {
-    return ASTPassResult(element: contractBehaviorMember, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(structDeclaration: StructDeclaration, passContext: ASTPassContext) -> ASTPassResult<StructDeclaration> {
-    return ASTPassResult(element: structDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(structMember: StructMember, passContext: ASTPassContext) -> ASTPassResult<StructMember> {
-    return ASTPassResult(element: structMember, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(enumDeclaration: EnumDeclaration, passContext: ASTPassContext) -> ASTPassResult<EnumDeclaration> {
-    return ASTPassResult(element: enumDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(enumCase: EnumCase, passContext: ASTPassContext) -> ASTPassResult<EnumCase> {
-    return ASTPassResult(element: enumCase, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(variableDeclaration: VariableDeclaration, passContext: ASTPassContext) -> ASTPassResult<VariableDeclaration> {
-    return ASTPassResult(element: variableDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(functionDeclaration: FunctionDeclaration, passContext: ASTPassContext) -> ASTPassResult<FunctionDeclaration> {
-    return ASTPassResult(element: functionDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(specialDeclaration: SpecialDeclaration, passContext: ASTPassContext) -> ASTPassResult<SpecialDeclaration> {
-    return ASTPassResult(element: specialDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(attribute: Attribute, passContext: ASTPassContext) -> ASTPassResult<Attribute> {
-    return ASTPassResult(element: attribute, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(parameter: Parameter, passContext: ASTPassContext) -> ASTPassResult<Parameter> {
-    return ASTPassResult(element: parameter, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(typeAnnotation: TypeAnnotation, passContext: ASTPassContext) -> ASTPassResult<TypeAnnotation> {
-    return ASTPassResult(element: typeAnnotation, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(identifier: Identifier, passContext: ASTPassContext) -> ASTPassResult<Identifier> {
-    return ASTPassResult(element: identifier, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(type: Type, passContext: ASTPassContext) -> ASTPassResult<Type> {
-    return ASTPassResult(element: type, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(callerCapability: CallerCapability, passContext: ASTPassContext) -> ASTPassResult<CallerCapability> {
-    return ASTPassResult(element: callerCapability, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(typeState: TypeState, passContext: ASTPassContext) -> ASTPassResult<TypeState> {
-    return ASTPassResult(element: typeState, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(expression: Expression, passContext: ASTPassContext) -> ASTPassResult<Expression> {
-    return ASTPassResult(element: expression, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(statement: Statement, passContext: ASTPassContext) -> ASTPassResult<Statement> {
-    return ASTPassResult(element: statement, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(inoutExpression: InoutExpression, passContext: ASTPassContext) -> ASTPassResult<InoutExpression> {
-    return ASTPassResult(element: inoutExpression, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(binaryExpression: BinaryExpression, passContext: ASTPassContext) -> ASTPassResult<BinaryExpression> {
-    return ASTPassResult(element: binaryExpression, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(functionCall: FunctionCall, passContext: ASTPassContext) -> ASTPassResult<FunctionCall> {
-    return ASTPassResult(element: functionCall, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(arrayLiteral: ArrayLiteral, passContext: ASTPassContext) -> ASTPassResult<ArrayLiteral> {
-    return ASTPassResult(element: arrayLiteral, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(rangeExpression: AST.RangeExpression, passContext: ASTPassContext) -> ASTPassResult<AST.RangeExpression> {
-    return ASTPassResult(element: rangeExpression, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(dictionaryLiteral: AST.DictionaryLiteral, passContext: AST.ASTPassContext) -> ASTPassResult<AST.DictionaryLiteral> {
-    return ASTPassResult(element: dictionaryLiteral, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(literalToken: Token, passContext: ASTPassContext) -> ASTPassResult<Token> {
-    return ASTPassResult(element: literalToken, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(subscriptExpression: SubscriptExpression, passContext: ASTPassContext) -> ASTPassResult<SubscriptExpression> {
-    return ASTPassResult(element: subscriptExpression, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(returnStatement: ReturnStatement, passContext: ASTPassContext) -> ASTPassResult<ReturnStatement> {
-    return ASTPassResult(element: returnStatement, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(becomeStatement: BecomeStatement, passContext: ASTPassContext) -> ASTPassResult<BecomeStatement> {
-    return ASTPassResult(element: becomeStatement, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(ifStatement: IfStatement, passContext: ASTPassContext) -> ASTPassResult<IfStatement> {
-    return ASTPassResult(element: ifStatement, diagnostics: [], passContext: passContext)
-  }
-
-  public func process(forStatement: ForStatement, passContext: ASTPassContext) -> ASTPassResult<ForStatement> {
-    return ASTPassResult(element: forStatement, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(topLevelModule: TopLevelModule, passContext: ASTPassContext) -> ASTPassResult<TopLevelModule> {
-    return ASTPassResult(element: topLevelModule, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(topLevelDeclaration: TopLevelDeclaration, passContext: ASTPassContext) -> ASTPassResult<TopLevelDeclaration> {
-    return ASTPassResult(element: topLevelDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(contractDeclaration: ContractDeclaration, passContext: ASTPassContext) -> ASTPassResult<ContractDeclaration> {
-    return ASTPassResult(element: contractDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(contractBehaviorDeclaration: ContractBehaviorDeclaration, passContext: ASTPassContext) -> ASTPassResult<ContractBehaviorDeclaration> {
-    return ASTPassResult(element: contractBehaviorDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(contractBehaviorMember: ContractBehaviorMember, passContext: ASTPassContext) -> ASTPassResult<ContractBehaviorMember> {
-    return ASTPassResult(element: contractBehaviorMember, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(structDeclaration: StructDeclaration, passContext: ASTPassContext) -> ASTPassResult<StructDeclaration> {
-    return ASTPassResult(element: structDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(structMember: StructMember, passContext: ASTPassContext) -> ASTPassResult<StructMember> {
-    return ASTPassResult(element: structMember, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(enumCase: EnumCase, passContext: ASTPassContext) -> ASTPassResult<EnumCase> {
-    return ASTPassResult(element: enumCase, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(enumDeclaration: EnumDeclaration, passContext: ASTPassContext) -> ASTPassResult<EnumDeclaration> {
-    return ASTPassResult(element: enumDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(variableDeclaration: VariableDeclaration, passContext: ASTPassContext) -> ASTPassResult<VariableDeclaration> {
-    return ASTPassResult(element: variableDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(functionDeclaration: FunctionDeclaration, passContext: ASTPassContext) -> ASTPassResult<FunctionDeclaration> {
-    return ASTPassResult(element: functionDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(specialDeclaration: SpecialDeclaration, passContext: ASTPassContext) -> ASTPassResult<SpecialDeclaration> {
-    return ASTPassResult(element: specialDeclaration, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(attribute: Attribute, passContext: ASTPassContext) -> ASTPassResult<Attribute> {
-    return ASTPassResult(element: attribute, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(parameter: Parameter, passContext: ASTPassContext) -> ASTPassResult<Parameter> {
-    return ASTPassResult(element: parameter, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(typeAnnotation: TypeAnnotation, passContext: ASTPassContext) -> ASTPassResult<TypeAnnotation> {
-    return ASTPassResult(element: typeAnnotation, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(identifier: Identifier, passContext: ASTPassContext) -> ASTPassResult<Identifier> {
-    return ASTPassResult(element: identifier, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(type: Type, passContext: ASTPassContext) -> ASTPassResult<Type> {
-    return ASTPassResult(element: type, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(callerCapability: CallerCapability, passContext: ASTPassContext) -> ASTPassResult<CallerCapability> {
-    return ASTPassResult(element: callerCapability, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(typeState: TypeState, passContext: ASTPassContext) -> ASTPassResult<TypeState> {
-    return ASTPassResult(element: typeState, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(expression: Expression, passContext: ASTPassContext) -> ASTPassResult<Expression> {
-    return ASTPassResult(element: expression, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(statement: Statement, passContext: ASTPassContext) -> ASTPassResult<Statement> {
-    return ASTPassResult(element: statement, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(inoutExpression: InoutExpression, passContext: ASTPassContext) -> ASTPassResult<InoutExpression> {
-    return ASTPassResult(element: inoutExpression, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(binaryExpression: BinaryExpression, passContext: ASTPassContext) -> ASTPassResult<BinaryExpression> {
-    return ASTPassResult(element: binaryExpression, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(functionCall: FunctionCall, passContext: ASTPassContext) -> ASTPassResult<FunctionCall> {
-    return ASTPassResult(element: functionCall, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(arrayLiteral: ArrayLiteral, passContext: ASTPassContext) -> ASTPassResult<ArrayLiteral> {
-    return ASTPassResult(element: arrayLiteral, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(rangeExpression: AST.RangeExpression, passContext: ASTPassContext) -> ASTPassResult<AST.RangeExpression> {
-    return ASTPassResult(element: rangeExpression, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(dictionaryLiteral: AST.DictionaryLiteral, passContext: ASTPassContext) -> ASTPassResult<AST.DictionaryLiteral> {
-    return ASTPassResult(element: dictionaryLiteral, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(literalToken: Token, passContext: ASTPassContext) -> ASTPassResult<Token> {
-    return ASTPassResult(element: literalToken, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(subscriptExpression: SubscriptExpression, passContext: ASTPassContext) -> ASTPassResult<SubscriptExpression> {
-    return ASTPassResult(element: subscriptExpression, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(returnStatement: ReturnStatement, passContext: ASTPassContext) -> ASTPassResult<ReturnStatement> {
-    return ASTPassResult(element: returnStatement, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(becomeStatement: BecomeStatement, passContext: ASTPassContext) -> ASTPassResult<BecomeStatement> {
-    return ASTPassResult(element: becomeStatement, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(ifStatement: IfStatement, passContext: ASTPassContext) -> ASTPassResult<IfStatement> {
-    return ASTPassResult(element: ifStatement, diagnostics: [], passContext: passContext)
-  }
-
-  public func postProcess(forStatement: ForStatement, passContext: ASTPassContext) -> ASTPassResult<ForStatement> {
-    return ASTPassResult(element: forStatement, diagnostics: [], passContext: passContext)
-  }
 }
